split repo.py and list_plugin.py
